+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2025-07-19
+
 ### Added
 
 - Intermediate upgrade progress event `UpgradedControlPlane` when control plane upgrade completes during cluster upgrades.
@@ -119,7 +121,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2024-08-13
 
-[Unreleased]: https://github.com/giantswarm/cluster-api-events/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-api-events/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/giantswarm/cluster-api-events/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/giantswarm/cluster-api-events/compare/v0.5.3...v0.6.0
 [0.5.3]: https://github.com/giantswarm/cluster-api-events/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/giantswarm/cluster-api-events/compare/v0.5.1...v0.5.2